Remove Setω from TT
First, TT is not actually using the fact that is defined in Setω. It
could just as well have been defined in, say, Set₀ and all examples
would still work.

Second, the definition of Map (and subsequent Fold, etc.) were highly
suspicious: they were significant size changes between sources and
targets. I'm not sure these are still related to the actual
categorical concepts.

This also manifests itself in the type of induction for LNDT, which is
uselessly generalized wrt. to the Set level of `A`.

Overall, having a down-to-earth definition for TT (and elsewhere) will
make my life easier.

(And premature universe level generalization is the root of all evil)