Iterative (equivalent) definition of LNDT
The idea here is to make it explicit that LNDT is about iterating the
functor `F : TT`. The traditional LNDT needlessly use an inductive
definition to encode both the depth of the structure (number of times
`F` gets iterated) as well as store its actual context.

This is but an historical accident: these things were born in simple
type theory. Being in type theory, we can separate these two aspects.